Carry Long through the JSON API pagination and pin it in the specs
DefaultJsonApiViewHelper read the pagination total as `(Integer)`, which
under static compilation is Groovy's numeric cast rather than a checkcast,
so a Long count wrapped silently — the same truncation this change removes
elsewhere, still reachable through jsonapi.render(). Read it as a Number
and widen with longValue().

PaginationSpec declared its results as Integer and only exercised int-range
values, so it passed by dynamic conversion without pinning the type it
tests. Declare them Long and add rows past Integer.MAX_VALUE, which also
cover the Math.round(Math.ceil(..)) to (long) Math.ceil(..) rewrite.
IterableRenderSpec gains a JSON API render with a total of 3_000_000_000,
asserting the last link keeps the 64-bit offset.

Widen offset to Long on the public paginate overloads so a page past 2^31
can be requested and not merely described. Reading it back needed
Parameters.long() rather than int(), or the request path would still have
capped at 2^31.

Upgrade note: describe the old pipeline rather than the new one, add the
JSON view model declaration alongside the GSP directive, and record that
paginate() and links() now declare total and offset as Long.

+++ grails-views-gson/src/main/groovy/grails/plugin/json/view/api/HalViewHelper.groovy
@@ -122,7 +122,7 @@ interface HalViewHelper {
      * @param total The total number of objects to be paginated
      * @param offset The numerical offset where the page starts (defaults to 0)
      */
-    void paginate(Object object, Long total, Integer offset)
+    void paginate(Object object, Long total, Long offset)
 
     /**
      * Pagination support which outputs hal links to the resulting pages
@@ -132,7 +132,7 @@ interface HalViewHelper {
      * @param offset The numerical offset where the page starts (defaults to 0)
      * @param max The maximum number of objects to be shown (defaults to 10)
      */
-    void paginate(Object object, Long total, Integer offset, Integer max)
+    void paginate(Object object, Long total, Long offset, Integer max)
 
     /**
      * Pagination support which outputs hal links to the resulting pages
@@ -143,7 +143,7 @@ interface HalViewHelper {
      * @param max The maximum number of objects to be shown (defaults to 10)
      * @param sort The field to sort on (defaults to null)
      */
-    void paginate(Object object, Long total, Integer offset, Integer max,  String sort)
+    void paginate(Object object, Long total, Long offset, Integer max,  String sort)
 
     /**
      * Pagination support which outputs hal links to the resulting pages
@@ -155,7 +155,7 @@ interface HalViewHelper {
      * @param sort The field to sort on (defaults to null)
      * @param order The order in which the results are to be sorted eg: DESC or ASC
      */
-    void paginate(Object object, Long total, Integer offset, Integer max,  String sort, String order)
+    void paginate(Object object, Long total, Long offset, Integer max,  String sort, String order)
 
     /**
      * Render embedded links for the associations of the given object (if any)
