+++ grails-views-gson/src/main/groovy/grails/plugin/json/view/api/internal/DefaultHalViewHelper.groovy
@@ -258,12 +258,12 @@ class DefaultHalViewHelper extends DefaultJsonViewHelper implements HalViewHelpe
      * @param order The order in which the results are to be sorted eg: DESC or ASC
      */
     //TODO: Once GROOVY-9662 is fixed, remove explicit delegate call and typecast to StreamingJsonDelegate
-    void paginate(Object object, Long total, Integer offset = null, Integer max = null,  String sort = null, String order = null) {
+    void paginate(Object object, Long total, Long offset = null, Integer max = null,  String sort = null, String order = null) {
         Map<String, Object> linkParams = buildPaginateParams(max, offset, sort, order)
 
         GrailsView jsonView = view
         Parameters httpParams = jsonView.params
-        offset = offset ?: httpParams.int(PAGINATION_OFFSET, 0)
+        offset = offset ?: httpParams.long(PAGINATION_OFFSET, 0L)
         max = max ?: httpParams.int(PAGINATION_MAX, 10)
         sort = sort ?: httpParams.get(PAGINATION_SORT)
         order = order ?: httpParams.get(PAGINATION_ORDER)
